Separated time control hooks into their own hook
