+++ src/App.tsx
@@ -1,16 +1,14 @@
 import React from "react";
 
 import { useConsentingAudio } from "./hooks/useConsentingAudio";
-import { useSettings } from "./useSettings";
 import { PlayView } from "./components/PlayView";
 
 const App: React.FC = () => {
-    const [settings, setSettings] = useSettings();
     const ConsentingAudioContext = useConsentingAudio();
 
     return (
         <ConsentingAudioContext>
-            <PlayView settings={settings} setSettings={setSettings} />
+            <PlayView />
         </ConsentingAudioContext>
     );
 };
